requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
CacheControl==0.14.3
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
firebase_admin==7.1.0
frozenlist==1.8.0
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
joblib==1.5.2
msgpack==1.1.2
multidict==6.7.0
numpy==2.3.3
pandas==2.3.3
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydub==0.25.1
pyfcm==2.1.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.3
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.22.0
